Change method of creating pandoc_params
Change method of defining s:dn_markdown_pandoc_params. Main issue is
that some formats derive their params (a List value) from another,
source, format. (That is, they have the same params List as the source
format.) The previous method of defining params was to set the params
value in the derived format to a string stub value. After the entire
variable was defined the source params Lists would be copied to the
derived format params Lists by individual copy commands.

The new method is to set the params value in derived formats to a Dict
that has a single key 'source' with the value set to the source format.
The dn#markdown#initialise() function calls s:_create_pandoc_params()
which processes the formats looking for params values set to a Dict.
When found, the source format is extracted and the source params List
copied to the derived params value.